Physics2D: Begin fixing Top-Level CollisionShapes
In the physics server, shapes are simple data structure without reference to the canvas, just a transform. This transform was always expected to be relative to the parent object's transform.

By adding a new field to these shapes (indpdt_xform) representing whether it is top_level, the final tranform computations can be adjusted to be relative or not